perf: do not search and dlopen on every sql cmd
This extracts a handle of zen-internal FFI to a thread safe singleton.
Ensuring the DLL is searched for and loaded only once. Previously, we
would run multiple syscalls to get the location of the dll/so and then
built the python datastructures (CDLL object) that has to be later GCed.

I haven't microbenchmarked the improvements rigorously, as I feel GC
savings may contribute considerably in real-world scenario.

Although,

    poetry run pytest -rP  ./aikido_zen/vulnerabilities/sql_injection/

improves by over 4.3% on my set-up.